Move BlockHash into more generic place (can be reused) and make endpoints for bitcoin more generic (does not require custom methods - to be verified)
