+++ conseil-api/src/main/scala/tech/cryptonomic/conseil/api/routes/platform/data/bitcoin/BitcoinDataOperations.scala
@@ -4,6 +4,7 @@ import slick.jdbc.PostgresProfile.api._
 import tech.cryptonomic.conseil.api.routes.platform.data.ApiDataOperations
 import tech.cryptonomic.conseil.common.bitcoin.Tables
 import tech.cryptonomic.conseil.common.bitcoin.Tables.{BlocksRow, TransactionsRow}
+import tech.cryptonomic.conseil.common.generic.chain.DataTypes.BlockHash
 import tech.cryptonomic.conseil.common.util.DatabaseUtil
 
 import scala.concurrent.Future
@@ -25,8 +26,8 @@ class BitcoinDataOperations extends ApiDataOperations {
     * @param hash The block's hash
     * @return The transaction, if the hash matches anything
     */
-  //TODO We should change this 'String' to 'BlockHash'
-  def fetchTransaction(hash: String): Future[Option[TransactionsRow]] =
-    runQuery(Tables.Transactions.filter(_.hash === hash).result.headOption)
+  //TODO Potentially not needed
+  def fetchTransaction(hash: BlockHash): Future[Option[TransactionsRow]] =
+    runQuery(Tables.Transactions.filter(_.hash === hash.value).result.headOption)
 
 }
