+++ conseil-api/src/main/scala/tech/cryptonomic/conseil/api/routes/platform/data/tezos/TezosDataOperations.scala
@@ -3,7 +3,8 @@ package tech.cryptonomic.conseil.api.routes.platform.data.tezos
 import com.github.ghik.silencer.silent
 import slick.jdbc.PostgresProfile.api._
 import tech.cryptonomic.conseil.api.routes.platform.data.ApiDataOperations
-import tech.cryptonomic.conseil.common.tezos.TezosTypes.{AccountId, BlockHash}
+import tech.cryptonomic.conseil.common.generic.chain.DataTypes.BlockHash
+import tech.cryptonomic.conseil.common.tezos.TezosTypes.AccountId
 import tech.cryptonomic.conseil.common.tezos.Tables
 import tech.cryptonomic.conseil.common.util.DatabaseUtil
 import tech.cryptonomic.conseil.common.util.CollectionOps._
@@ -128,6 +129,9 @@ class TezosDataOperations extends ApiDataOperations {
     * @return the most recent block, if one exists in the database.
     */
   def latestBlockIO()(implicit ec: ExecutionContext): DBIO[Option[Tables.BlocksRow]] =
+  // Couldn't this query be done in single database request?
+  // I believe we could simply sort + take. Wouldn't this be faster?
+  // Probably we should test this out first
     fetchMaxBlockLevel.flatMap(
       maxLevel =>
         Tables.Blocks
