+++ conseil-api/src/test/scala/tech/cryptonomic/conseil/api/routes/platform/data/tezos/TezosDataOperationsTest.scala
@@ -18,7 +18,7 @@ import tech.cryptonomic.conseil.common.tezos.Tables.{
   OperationsRow
 }
 import tech.cryptonomic.conseil.common.tezos.Tables
-import tech.cryptonomic.conseil.common.tezos.TezosTypes.{AccountId, BlockHash}
+import tech.cryptonomic.conseil.common.tezos.TezosTypes.AccountId
 
 import scala.concurrent.duration._
