+++ conseil-lorre/src/main/scala/tech/cryptonomic/conseil/indexer/tezos/TezosDatabaseOperations.scala
@@ -10,7 +10,7 @@ import org.slf4j.LoggerFactory
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.{AbstractTable, TableQuery}
 import tech.cryptonomic.conseil.common.config.ChainEvent.AccountIdPattern
-import tech.cryptonomic.conseil.common.generic.chain.DataTypes.{Query => _}
+import tech.cryptonomic.conseil.common.generic.chain.DataTypes.{BlockHash, Query => _}
 import tech.cryptonomic.conseil.common.sql.CustomProfileExtension
 import tech.cryptonomic.conseil.common.tezos.Tables.{GovernanceRow, OriginatedAccountMapsRow}
 import tech.cryptonomic.conseil.common.tezos.TezosTypes.Fee.AverageFees
@@ -24,6 +24,7 @@ import tech.cryptonomic.conseil.common.util.CollectionOps._
 import tech.cryptonomic.conseil.common.util.Conversion.Syntax._
 import tech.cryptonomic.conseil.common.util.MathUtil.{mean, stdev}
 import tech.cryptonomic.conseil.indexer.sql.DefaultDatabaseOperations._
+
 import scala.collection.immutable.Queue
 import scala.concurrent.{ExecutionContext, Future}
 import scala.math.{ceil, max}
