+++ conseil-lorre/src/test/scala/tech/cryptonomic/conseil/indexer/tezos/TezosNodeOperatorTest.scala
@@ -4,8 +4,8 @@ import com.typesafe.scalalogging.LazyLogging
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.{FlatSpec, Matchers}
+import tech.cryptonomic.conseil.common.generic.chain.DataTypes.BlockHash
 import tech.cryptonomic.conseil.common.tezos.TezosTypes
-import tech.cryptonomic.conseil.common.tezos.TezosTypes.BlockHash
 import tech.cryptonomic.conseil.indexer.config.BatchFetchConfiguration
 
 import scala.concurrent.duration._
